requirements.txt
Flask
flask-socketio
eventlet
python-socketio
python-engineio
pydantic
msgpack
pyserial